+++ NEWS.md
@@ -1,3 +1,11 @@
+# jrt 1.0.1
+
+* Simulated dataset is now directly included as `jrt::ratings`.
+* Model weights (often called Akaike weights) are now presented in the model selection output (per reviewer suggestion).
+* Calls to `mirt` can now be showed with `show.calls = TRUE` argument in the `jrt()`function.
+* References were updated.
+
+
 # jrt 1.0.0
 
 * Initial release with functions jrt(), jcc.plot() and info.plot()